Devided some functions into utils.js
- Devided some functions into utils.js
